+++ Changelog.md
@@ -8,3 +8,4 @@
 - Using regexps, ranges and arrays within `where` clause
 - `where.not(...)` support
 - Add `past` and `since` methods
+- Add `merge` method and support for regexp series